Correctly deal with missing HTTP Headers
This patch provides a complete fix to issues that occur when UCL's shibboleth
instance does not return certain headers.

The following fields are considered critical and cannot be empty. Absence of
these parameters will result in a 400 error code being returned:
	1. eppn - used for finding user in DB. (We need to reconsider this...)
	2. cn - unique in DB
	3. employee_id - unique in DB, and used for determining is_student

The following fields are considered non-compulsory and instead replaced with an
empty string:
	1. department
	2. given_name
	3. display_name
	4. groups

This patch updates the documentation accordingly, and includes tests for certain
scenarios.